fix: dont freeze editor when closing editor with offline cluster
Freezing was caused by race conditions in the locks. Closing the editor
disposes the editor resources. Disposing the editor resource closes the
cluster connection. When closing the watch, a new resource state is
retrieved, this then ended in the resource lock waiting for the
connection timeout.
Solved it by introducing a new "Disposed" state which is returned
whenever an editor resource is disposed. No further state retrieval
occurrs.

Signed-off-by: Andre Dietisheim <adietish@redhat.com>